backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "peewee>=4.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
    "scipy>=1.17.1",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.34.0",
    "xai-sdk>=1.7.0",
    "twilio>=9.10.2",
]

[dependency-groups]
dev = []
